fix: Make sure to add chunks that are nowhere referenced
If a rollup chunk is not referenced in any edges, it ended up not being added to the set of "leafs".
This fixes the issue. This problem was not prominent since such cases are unnatural -- in practice,
any dependencies among output modules are specified because of actual need, that is, actual sharing of
codes. However, if one specifies dependencies for modules which does not share any code this problem can
occur.