refactor: enable "strict" TS compiler option repository-wise
Initially suggested at https://github.com/theseanl/tscc/pull/434.

Maybe in one day one has to maintain a fork of tsickle in this
repository. In such situation, it'd be better to have a single tsconfig
that applies to the whole package. There were also a couple of places
where this may have changed runtime behavior in some edge cases.

However, this added a lot of nuisance to the previously-clean codebase:
Map-like class implementations were especially not suited for the strict
null checks. In lots of places, we do `if (this.map.has(key)) { do
something with this.map.get(key)` } and TS still thinks that the return
value of `get` may be null, and there seems to be no type guards
available for this patern.