fix: exclude dts shipped with tscc from compilation
The use case for base.d.ts and reflect.d.ts is for codes referencing
compiler primitives such as `goog.define` and `goog.reflect`. Such codes
can reference the relevant `.d.ts` files via triple-slash directives
while not including them in compilation. This behavior has been bugged
since we changed to use TypescriptDependencyGraph approach, and this
commit fixes it.

In addition, now `base.d.ts` exposes `goog` as a global variable. It
appears that one cannot `goog.require('goog')`, and in tsickle it is
mentioned that within google one does `import * as goog from
'google3.javascript.closure.goog'`, and tsickle has a special casing for
this. This is not available, hence the approach we have taken.