feat: fix escape_goog_identifier to not emit consecutive periods
Previously, it may emit consecutive periods for input "./". We expected
such a string to never be provided to the function, because all files
should be located under the project root. However, this change wouldn't
hurt, and we may revisit this assumption later. Also, there could be
file names that ends with periods, however unlikely it is.