feat: update tsickle to 0.46.3; typescript to 4.7
It became clear that tsickle does not support targetting ES3. Thus, tscc
will now print a warning if it is provided with a tsconfig targetting ES3,
and will override it to ES5. I have considered setting compilerFlag
language_out=ECMASCRIPT_3 in such cases, but I went down the path of
informing users to set it manually, because of a need of refactoring
just for this case.

An internal API ts.SourceFile.resolvedModules has been changed from ES6
Map to a custom map-like object, so an appropriate change had to be made
to TypescriptDependencyGraph.ts.