feat: implement es6 chunk generation
This commit adds a new feature that allows users to perform build that
generates es6 modules as output chunks. Previously, output chunks need
to be executed in a shared global scope in order for later chunks to
reference variables in previous chunks. By adding `chunkFormat` key to
the spec file, the closure compiler build will be done using
chunk_output_type = ES_MODULES, and rollup build will be done using a
usual code-splitting build with tight control of each chunk's location
in the final bundle.

Currently it does not work with the spec file's `external` option. It
"sort of" works for rollup, but again it still doesn't work with
external modules specified with relative paths. On the closure compiler
side, it is even less clear what should be done, but it seems clear that
this is the way to go.